pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-interaction"
version = "0.1.0"
description = "Async Azure OpenAI client with typed tool calling and lazy output parsing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "jinja2>=3.0",
    "json-repair>=0.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_interaction"]
